Moved Logo to fortanix directory
