+++ src/fortanix/Logo/Logo.module.scss
@@ -5,7 +5,7 @@
 /* stylelint-disable no-duplicate-selectors */
 /* stylelint-disable selector-class-pattern */
 
-@use '../../../styling/defs.scss' as bk;
+@use '../../styling/defs.scss' as bk;
 
 @layer baklava.components {
   .bk-app-logo {
