+++ src/fortanix/Logo/Logo.tsx
@@ -2,10 +2,10 @@
 |* This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. If a copy of
 |* the MPL was not distributed with this file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { classNames as cx, type ComponentProps } from '../../../util/componentUtil.ts';
+import { classNames as cx, type ComponentProps } from '../../util/componentUtil';
 import * as React from 'react';
 
-import logoImage from '../../../assets/fortanix/fortanix_logo_icon.svg';
+import logoImage from  '../../assets/fortanix/fortanix_logo_icon.svg';
 import cl from './Logo.module.scss';
